Move all functionality into App class
This allows us to get rid of all global state (although in a pretty
heavy-handed way), and makes it easy to continue refactoring
without having to care too much about dependencies until
we get to a better structure.

I've not cleaned up any of the code after
placing it into the class.